Отрисовывает жанры отдельно от главного экрана
