+++ project/src/pages/movie-page/movie-page.tsx
@@ -10,6 +10,7 @@ import { AuthorizationStatus } from '../../const';
 import Tabs from '../../components/tabs/tabs';
 import { getFilm, getReviews, getSimilarFilms } from '../../store/film-data/selectors';
 import { getAuthorizationStatus } from '../../store/user-process/selectors';
+import { getFilms } from '../../store/films-data/selectors';
 
 export default function MoviePage(): JSX.Element {
   const params = useParams();
@@ -21,6 +22,8 @@ export default function MoviePage(): JSX.Element {
   const similarFilms = useAppSelector(getSimilarFilms);
   const authorizationStatus = useAppSelector(getAuthorizationStatus);
   const reviews = useAppSelector(getReviews);
+  const films = useAppSelector(getFilms);
+  const favoriteFilms = films.filter((film) => film.isFavorite === true).length;
 
   useEffect(() => {
     if (id === null) {
@@ -76,7 +79,7 @@ export default function MoviePage(): JSX.Element {
                     <use xlinkHref="#add"></use>
                   </svg>
                   <span>My list</span>
-                  <span className="film-card__count">9</span>
+                  <span className="film-card__count">{favoriteFilms}</span>
                 </button>
                 {(authorizationStatus === AuthorizationStatus.Auth)
                   ? <Link to={`/films/${currentFilm?.id}/review`} className="btn film-card__button">Add review</Link>
